+++ ApiGateway/villager/get_detail/schema.json
