+++ aws_s3_bucket.tf
@@ -1,8 +1,3 @@
 resource "aws_s3_bucket" "s3" {
-  bucket = "${var.app}-codebuild"
-}
-
-resource "aws_s3_bucket_policy" "s3" {
-  bucket = aws_s3_bucket.s3.id
-  policy = data.aws_iam_policy_document.s3_bucket.json
+  bucket = "${var.app}-${var.aws_region}-${var.aws_account_id}"
 }
