+++ aws_s3_bucket_versioning.tf
@@ -1,0 +1,6 @@
+resource "aws_s3_bucket_versioning" "s3" {
+  bucket = aws_s3_bucket.s3.id
+  versioning_configuration {
+    status = "Disabled"
+  }
+}
