+++ aws_secretsmanager_secret.tf
@@ -1,12 +1,16 @@
-
+/*
 resource "null_resource" "delete_aws_secretsmanager_secret_rds" {
-  count      = terraform.workspace == "destroy" ? 1 : 0
   depends_on = [aws_secretsmanager_secret.rds]
 
   provisioner "local-exec" {
     command = "aws secretsmanager delete-secret --secret-id ${aws_secretsmanager_secret.rds.name} --force-delete-without-recovery"
   }
+  
+  triggers = {
+    destroy_secret_id = aws_secretsmanager_secret.rds.id
+  }
 }
+*/
 
 resource "aws_secretsmanager_secret" "rds" {
   name = "${aws_db_instance.animal_crossing.identifier}-rds"
